+++ lib/nugrant/parameters.rb
@@ -15,17 +15,11 @@ module Nugrant
     #
     # =| Arguments
     #  * `config`
-    #    A hash that will be passed to Nugrant::Config.new() or
-    #    a Nugrant::Config instance directly.
-    #    See Nugrant::Config constructor for options that you can use.
+    #    Passed to Mixin::Parameters setup! method. See method
+    #    for further information.
     #
-    #  * `options`
-    #    An options hash that is passed to Mixin::Parameters.compute_bags! method.
-    #    See Mixin::Parameters.compute_bags! for details on the various options
-    #    available.
-    #
-    def initialize(config, options = {})
-      compute_bags!(config, options)
+    def initialize(config)
+      setup!({}, config)
     end
 
     include Mixin::Parameters